Move AWS worker to its own package
