+++ internal/worker/aws/generate_test.go
@@ -1,0 +1,3 @@
+package aws
+
+//go:generate mockgen -source=aws.go -destination=aws_mock.go -package aws
