+++ internal/worker/aws/userdata.sh
